+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [2.1.0]
 
 ### Additions
+- [Issue #11] onComplete should accept functions that return Promise
+- [Issue #10] BootstrapFunction should accept return type void or Promise<void>
 - [Issue #7] Add onFinally method
 
 ### Fixes